requirements.txt
mtcnn==1.0.0
keras-facenet==0.3.2
tensorflow == 2.18.0
keras==3.8.0
opencv-python==4.11.0.86
numpy==2.0.2
scipy==1.14.1
qdrant-client==1.13.3
pytest==8.3.5
dotenv==0.9.9
pillow==11.1.0
fpdf==1.7.2
qrcode==8.0.0
flask==3.1.0
supabase==2.13.0
pyOpenSSL==24.0.0
cryptography==42.0.0
flask-wtf==1.2.2
PyJWT==2.10.0
dlib==19.24.6
google-generativeai==0.6.15
PyPDF2==3.0.1